refactor: ResultStatus 네이밍 변경 -> GameResult, 이중 스트림 메서드 분리
