+++ src/test/java/blackjack/domain/result/GameResultTest.java
@@ -6,7 +6,6 @@ import blackjack.domain.card.Face;
 import blackjack.domain.card.Suit;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Player;
-import blackjack.domain.participant.Players;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -41,7 +40,7 @@ public class ResultStatisticsTest {
         Player player = new Player(new Cards(cardList), "pobi");
         players.add(player);
 
-        ResultStatistics resultStatistics = new ResultStatistics(new Players(players), dealer);
+        ResultStatistics resultStatistics = new ResultStatistics(players, dealer);
 
         assertThat(resultStatistics.getDealerLoseCounts()).isEqualTo(1);
     }
@@ -57,7 +56,7 @@ public class ResultStatisticsTest {
         Player player = new Player(new Cards(cardList), "pobi");
         players.add(player);
 
-        ResultStatistics resultStatistics = new ResultStatistics(new Players(players), dealer);
+        ResultStatistics resultStatistics = new ResultStatistics(players, dealer);
 
         assertThat(resultStatistics.getDealerDrawCounts()).isEqualTo(1);
 
@@ -74,7 +73,7 @@ public class ResultStatisticsTest {
         Player player = new Player(new Cards(cardList), "pobi");
         players.add(player);
 
-        ResultStatistics resultStatistics = new ResultStatistics(new Players(players), dealer);
+        ResultStatistics resultStatistics = new ResultStatistics(players, dealer);
 
         assertThat(resultStatistics.getDealerWinCounts()).isEqualTo(1);
     }
@@ -91,7 +90,7 @@ public class ResultStatisticsTest {
         Player player = new Player(new Cards(cardList), "pobi");
         players.add(player);
 
-        ResultStatistics resultStatistics = new ResultStatistics(new Players(players), dealer);
+        ResultStatistics resultStatistics = new ResultStatistics(players, dealer);
 
         assertThat(resultStatistics.getDealerWinCounts()).isEqualTo(1);
     }
@@ -112,7 +111,7 @@ public class ResultStatisticsTest {
         cards.add(new Card(Suit.HEART, Face.JACK));
 
         Dealer dealer1 = new Dealer(cards);
-        ResultStatistics resultStatistics = new ResultStatistics(new Players(players), dealer1);
+        ResultStatistics resultStatistics = new ResultStatistics(players, dealer1);
 
         assertThat(resultStatistics.getDealerWinCounts()).isEqualTo(1);
     }
